requirements.txt
# ============================================================
# JARVIS AI ASSISTANT — DEPENDENCIES
# ============================================================
# Install: pip install -r requirements.txt
# ============================================================

# --- Core AI Framework ---
openai-agents>=0.14.0
openai>=1.0.0

# --- LLM Providers ---
python-dotenv>=1.0.0

# --- Voice: Speech-to-Text ---
faster-whisper>=1.0.0
sounddevice>=0.4.0
numpy>=1.24.0
scipy>=1.10.0

# --- Voice: Text-to-Speech ---
edge-tts>=6.0.0
pygame>=2.5.0

# --- UI & Formatting ---
colorama>=0.4.6
rich>=13.0.0

# --- Utilities ---
pydantic>=2.0.0
aiohttp>=3.9.0
aiofiles>=23.0.0
